Finish initial draft of meetup-events endpoint
