+++ functions/meetup-events/package.json
@@ -1,14 +1,7 @@
 {
   "name": "meetup-events",
-  "version": "1.0.0",
-  "description": "",
+  "private": true,
   "main": "meetup-events.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "dependencies": {
     "node-fetch": "^2.6.0"
   }
